OBJweb/requirements.txt
flask
paho-mqtt
numpy
opencv-python